services/notification/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
alembic==1.13.1
pytest==7.4.4
httpx==0.26.0
python-jose[cryptography]
passlib[bcrypt]
pytest-asyncio==0.23.3
